+++ MPMG.Repositories/Entidades/CupomFiscal.cs
@@ -8,10 +8,10 @@ namespace MPMG.Repositories.Entidades
         public string Coo { get; set; }
         public string NumeroNotaFiscal { get; set; }
         public string PostoReferente { get; set; }
-        public int Hodometro { get; set; }
+        public double Hodometro { get; set; }
         public string Cliente { get; set; }
         public DateTime DataEmissao { get; set; }
-        public int Quantidade { get; set; }
+        public double Quantidade { get; set; }
         public double PrecoUnitario { get; set; }
         public double ValorTotal { get; set; }
         public string Produto { get; set; }
